tools: keep which database the tables read apart from what is read of it
`tools/unicode/case_data.rb` said two things: which release of the Unicode
Character Database the tables are generated from, with the digest of each
file, and what the case mappings are once the files are read. The first is
not about case. A generator reading the database for something else would
have to name the same release and check the same digests, and could only do
so by requiring the case reader.

Move the release, the files, their digests, where they are and how they are
checked into `tools/unicode/ucd.rb`, and leave `case_data.rb` the reading.
Every file is checked whichever table is being generated, since the tables
are regenerated together and are to read one release between them. Nothing
generated changes: `rake unicode:verify` finds the committed tables up to
date.
